--- components/articles/article-tranches/bouton-supprimer-tranche.module.scss
@@ -1,5 +1,0 @@
-@import 'styles/variables.scss';
-
-.amendementText {
-  color: $reform-text-color !important;
-}
