+++ components/articles/articles-component.jsx
@@ -1,4 +1,6 @@
 import Grid from "@material-ui/core/Grid";
+import AddIcon from "@material-ui/icons/Add";
+import DeleteIcon from "@material-ui/icons/Delete";
 import classNames from "classnames";
 import PropTypes from "prop-types";
 import React, { Fragment } from "react";
@@ -13,24 +15,10 @@ import Alinea2 from "./article-alinea-2";
 import Alinea3 from "./article-alinea-3";
 import Alinea4a from "./article-alinea-4a";
 import ArticleHeader from "./article-header";
-import BoutonAjouterTranche from "./article-tranches/bouton-ajouter-tranche";
-import BoutonSupprimerTranche from "./article-tranches/bouton-supprimer-tranche";
 import styles from "./articles.module.scss";
+import { Button } from "./buttons";
 import { PrimaryExpandablePanel, SecondaryExpandablePanel, TertiaryExpandablePanel } from "./expandable-panels";
 import { ReglesGenerales, ReglesSpecifiques } from "./quotient-familial";
-
-const style = {
-  Button: {
-    margin: "10px",
-    padding: "3px",
-  },
-  Div: {
-    marginBottom: "1.5em",
-    marginLeft: "1.5em",
-    marginRight: "1.5em",
-    marginTop: "0",
-  },
-};
 
 class ArticlesComponent extends React.Component {
   // const nbt = props.reformeBase.impot_revenu.bareme.seuils.length;
@@ -195,13 +183,10 @@ class ArticlesComponent extends React.Component {
             {articleTranches}
             <Grid container spacing={0}>
               <Grid item sm={3}>
-                <BoutonAjouterTranche style={style} onClick={handleAddTranche} />
+                <Button caption="Ajouter ou" icons={<AddIcon />} onClick={handleAddTranche} />
               </Grid>
               <Grid item sm={6}>
-                <BoutonSupprimerTranche
-                  style={style}
-                  onClick={handleRemoveTranche}
-                />
+                <Button caption="Supprimer une tranche" icons={<DeleteIcon />} onClick={handleRemoveTranche} />
               </Grid>
             </Grid>
           </PrimaryExpandablePanel>
