+++ components/articles/buttons/Button.module.scss
@@ -2,4 +2,5 @@
 
 .amendementText {
   color: $reform-text-color !important;
+  cursor: pointer;
 }
